Se arreglo el fallo de seguridad de un usuario sin permiso

+++ application/views/admin/categorias/list.php
@@ -16,7 +16,9 @@
             <div class="box-body">
                 <div class="row">
                     <div class="col-md-12">
-                        <a href="<?php echo base_url();?>mantenimiento/categorias/add" class="btn btn-primary btn-flat"><span class="fa fa-plus"></span> Agregar Categoria</a>
+                        <?php if($permisos->insert == 1): ?>
+                            <a href="<?php echo base_url();?>mantenimiento/categorias/add" class="btn btn-primary btn-flat"><span class="fa fa-plus"></span> Agregar Categoria</a>
+                        <?php endif; ?>
                     </div>
                 </div>
 
@@ -45,8 +47,14 @@
                                             <button type="button" class="btn btn-info btn-view" data-toggle="modal" data-target="#modal-default" value="<?php echo $categoria->id;?>">
                                               <span class="fa fa-search"></span>
                                             </button>
-                                            <a href="<?php echo base_url();?>mantenimiento/categorias/edit/<?php echo $categoria->id;?>" class="btn btn-warning"><span class="fa fa-pencil"></span></a>
-                                            <a href="<?php echo base_url();?>mantenimiento/categorias/delete/<?php echo $categoria->id; ?>" class="btn btn-danger btn-remove"><span class="fa fa-remove"></span></a>
+                                            <!--Consulta si existe el permiso para editar-->
+                                            <?php if($permisos->update == 1): ?>
+                                                <a href="<?php echo base_url();?>mantenimiento/categorias/edit/<?php echo $categoria->id;?>" class="btn btn-warning"><span class="fa fa-pencil"></span></a>
+                                            <?php endif; ?>
+                                            <!--Consulta si existe el permiso para eliminar-->
+                                            <?php if($permisos->delete == 1): ?>
+                                                <a href="<?php echo base_url();?>mantenimiento/categorias/delete/<?php echo $categoria->id; ?>" class="btn btn-danger btn-remove"><span class="fa fa-remove"></span></a>
+                                            <?php endif; ?>
                                         </div>
                                     </td>
                                 </tr>
